[backend][security] Security Implemented

+++ FINTed_android/app/src/main/java/it/unical/demacs/enterprise/fintedapp/viewmodels/OfferViewModel.kt.kt
@@ -9,7 +9,9 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 
-class OfferViewModel: ViewModel() {
+class OfferViewModel(userViewModel: MutableState<UserViewModel>): ViewModel() {
+    private val userViewModel = userViewModel
+
     private val offerApi: OfferControllerApi = OfferControllerApi()
 
     val offerList: MutableState<List<OfferDto>> = mutableStateOf(listOf())
@@ -18,44 +20,52 @@ class OfferViewModel: ViewModel() {
 
     fun save(postId: Long, userId: Long, offer: String) {
         CoroutineScope(Dispatchers.IO).launch {
-            offerApi.save3(
-                OfferDto(
-                    postId = postId,
-                    userId = userId,
-                    offer =  offer.toLong()
+            userViewModel.value.accessTokenResponse.value.access_token?.let {
+                offerApi.save3(
+                    OfferDto(
+                        postId = postId,
+                        userId = userId,
+                        offer =  offer.toLong(),
+                        userUsername = userViewModel.value.personalProfile.value.username
+                    ),
+                    token = it
                 )
-            )
+            }
         }
     }
 
-    fun delete(offerId: Long){
+    fun delete(offerId: Long, username: String){
         CoroutineScope(Dispatchers.IO).launch {
-            offerApi.delete3(offerId)
+            userViewModel.value.accessTokenResponse.value.access_token?.let { offerApi.delete3(offerId, username, token = it) }
         }
     }
 
     fun getPostOffers(postId: Long){
         CoroutineScope(Dispatchers.IO).launch {
-            offerList.value = offerApi.getPostOffers(postId).toList()
+            offerList.value = userViewModel.value.accessTokenResponse.value.access_token?.let {
+                offerApi.getPostOffers(postId,
+                    it
+                ).toList()
+            }!!
         }
     }
 
     fun getUserOffers(userId: Long){
         CoroutineScope(Dispatchers.IO).launch {
-            offerList.value = offerApi.getUserOffers(userId).toList()
+            offerList.value = userViewModel.value.accessTokenResponse.value.access_token?.let { offerApi.getUserOffers(userId, token = it).toList() }!!
         }
     }
 
     fun getSellOffers(userId: Long){
         CoroutineScope(Dispatchers.IO).launch {
-            offerList.value = offerApi.getSellOffers(userId).toList()
+            offerList.value = userViewModel.value.accessTokenResponse.value.access_token?.let { offerApi.getSellOffers(userId, token = it).toList() }!!
         }
     }
 
     fun acceptOffer(offerDto: OfferDto): Boolean{
         var accepted = false;
         CoroutineScope(Dispatchers.IO).launch {
-            offer.value = offerApi.acceptOffer(offerDto)
+            offer.value = userViewModel.value.accessTokenResponse.value.access_token?.let { offerApi.acceptOffer(offerDto, token = it) }!!
             if(offer.value.offerStatus == OfferDto.OfferStatus.ACCEPTED){
                 accepted = true
             }
@@ -67,7 +77,7 @@ class OfferViewModel: ViewModel() {
     fun denyOffer(offerDto: OfferDto): Boolean{
         var denied = false;
         CoroutineScope(Dispatchers.IO).launch {
-            offer.value = offerApi.denyOffer(offerDto)
+            offer.value = userViewModel.value.accessTokenResponse.value.access_token?.let { offerApi.denyOffer(offerDto, token = it) }!!
             if(offer.value.offerStatus == OfferDto.OfferStatus.DENIED){
                 denied = true
             }
